# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Agentic SDLC Platform API
  description: REST API for the multi-agent SDLC orchestration platform.
  version: 0.4.0
...
paths:
...
  /api/v1/approvals/{id}/detail:
    get:
      tags:
        - cockpit
      parameters:
        - schema:
            type: string
          in: path
          name: id
          required: true
      responses:
        "200":
          description: Default Response
...
servers:
  - url: http://localhost:3000
    description: Local development server
...